Add unique data for Chosen playbook
I wrote the unique data for the Chosen playbook and added it to the JSON
file. I wrote tests to confirm this data is being added in the correct format.
I also added an expect() call in the tests that confirm the URL for a
tag is making a successful call to
/api/v1/weapon-tag/:index.